feat: Update GuestLayout, FrontendController, CategorySeeder, HeroSection, and NavbarView
This commit updates the GuestLayout component to add a top margin of 20 pixels to the FooterView. It also renames the `experiment` method in the FrontendController to `menu`. The CategorySeeder is modified to change the name of a category from 'Makanan Berat' to 'Catering'. The HeroSection component now has a top margin of 16 pixels. Lastly, the NavbarView component is updated to change the path for the 'Menu' navigation item to '/menu'. These changes improve the layout, functionality, and navigation of the frontend pages.